# api.yaml
# Worker API 接口说明
# Base URL: http://{host}:{port} (默认端口 8080)

openapi: "3.0.0"
info:
  title: "Test Worker API"
  description: "多端自动化测试执行基建 API"
  version: "2.0.0"

servers:
  - url: "http://{host}:{port}"
    variables:
      host:
        default: "localhost"
      port:
        default: "8080"

paths:
  /health:
    get:
      summary: "健康检查"
      description: "检查 Worker 服务是否正常运行"
      responses:
        "200":
          description: "服务正常"
          content:
            application/json:
              example:
                status: "ok"
                timestamp: "2026-03-14T10:30:00.000000"

  /status:
    get:
      summary: "获取 Worker 状态"
      description: "获取 Worker 详细状态信息"
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              example:
                worker_id: "worker_abc123"
                status: "running"
                started_at: "2026-03-14T10:00:00.000000"
                supported_platforms: ["web", "windows", "android", "ios"]
                active_sessions: 2
                devices_count: 3
        "503":
          description: "Worker 未初始化"

  /info:
    get:
      summary: "获取 Worker 详细信息"
      description: "获取支持的平台、设备列表等完整信息"
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              example:
                worker_id: "worker_abc123"
                hostname: "DESKTOP-ABC"
                ip_addresses: ["192.168.1.100", "10.0.0.1"]
                os_type: "Windows"
                os_version: "10.0.19045"
                supported_platforms: ["web", "windows", "android", "ios"]
                status: "running"
                port: 8080
                devices:
                  - id: "emulator-5554"
                    platform: "android"
                    name: "Pixel_4_API_30"

  /task/execute:
    post:
      summary: "执行任务（同步）"
      description: "同步执行自动化测试任务，阻塞等待结果返回"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
            example:
              platform: "web"
              actions:
                - action_type: "navigate"
                  value: "https://example.com"
                - action_type: "ocr_click"
                  value: "登录"
                - action_type: "ocr_input"
                  value: "用户名"
                  offset:
                    x: 100
                    y: 0
                - action_type: "screenshot"
                  value: "result"
              device_id: null
              user_id: "user_001"
              session_id: null
              config:
                timeout: 300000
                action_timeout: 30000
                screenshot_on_error: true
                slow_motion: 0
                retry_count: 0
              callback_url: null
              priority: 0
      responses:
        "200":
          description: "任务执行完成"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResult"
        "503":
          description: "Worker 未初始化"

  /task:
    post:
      summary: "提交任务（异步）"
      description: "提交任务到队列，异步执行"
      responses:
        "501":
          description: "功能未实现"

  /task/{task_id}:
    get:
      summary: "查询任务结果"
      description: "根据任务 ID 查询执行结果"
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "501":
          description: "功能未实现"

  /task/{task_id}:
    delete:
      summary: "取消任务"
      description: "取消正在执行的任务"
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "501":
          description: "功能未实现"

  /session:
    post:
      summary: "创建会话"
      description: "创建一个新的自动化会话，可用于复用浏览器/应用实例"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionRequest"
            example:
              platform: "web"
              device_id: null
              options:
                headless: false
                browser_type: "chromium"
      responses:
        "200":
          description: "会话创建成功"
          content:
            application/json:
              example:
                session_id: "sess_xyz789"
                platform: "web"
                device_id: null
                created_at: "2026-03-14T10:30:00.000000"
        "500":
          description: "创建失败"

  /session/{session_id}:
    get:
      summary: "获取会话状态"
      description: "查询指定会话的状态信息"
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
        - name: platform
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              example:
                session_id: "sess_xyz789"
                platform: "web"
                device_id: null
                created_at: "2026-03-14T10:30:00.000000"
                last_active: "2026-03-14T10:35:00.000000"
        "404":
          description: "会话不存在"

    delete:
      summary: "关闭会话"
      description: "关闭并销毁指定会话"
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
        - name: platform
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              example:
                closed: true

  /devices:
    get:
      summary: "获取设备列表"
      description: "获取当前连接的所有设备信息"
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              example:
                - id: "emulator-5554"
                  platform: "android"
                  name: "Pixel_4_API_30"
                - id: "00008030-001234567890ABCD"
                  platform: "ios"
                  name: "iPhone 14"

  /devices/refresh:
    post:
      summary: "刷新设备列表"
      description: "重新扫描并更新设备列表"
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              example:
                devices:
                  - id: "emulator-5554"
                    platform: "android"
                    name: "Pixel_4_API_30"
                added: ["new_device_id"]
                removed: ["old_device_id"]

  /screenshot:
    post:
      summary: "获取实时截图"
      description: "获取指定会话的当前屏幕截图"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreenshotRequest"
            example:
              platform: "web"
              session_id: "sess_xyz789"
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              example:
                image: "iVBORw0KGgoAAAANSUhEUgAA..."
                format: "png"
        "500":
          description: "获取失败"

components:
  schemas:
    TaskRequest:
      type: object
      required:
        - platform
        - actions
      properties:
        platform:
          type: string
          enum: [web, android, ios, windows, mac]
          description: "目标平台"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/Action"
          description: "动作列表"
        device_id:
          type: string
          description: "设备 ID（移动端必填）"
        user_id:
          type: string
          description: "用户标识"
        session_id:
          type: string
          description: "会话 ID（复用会话）"
        config:
          $ref: "#/components/schemas/TaskConfig"
        callback_url:
          type: string
          description: "回调地址"
        priority:
          type: integer
          default: 0
          description: "优先级"

    TaskConfig:
      type: object
      properties:
        timeout:
          type: integer
          default: 300000
          description: "总超时时间(ms)"
        action_timeout:
          type: integer
          default: 30000
          description: "单个动作超时时间(ms)"
        screenshot_on_error:
          type: boolean
          default: true
          description: "错误时是否截图"
        slow_motion:
          type: integer
          default: 0
          description: "慢动作延迟(ms)"
        retry_count:
          type: integer
          default: 0
          description: "失败重试次数"

    Action:
      type: object
      required:
        - action_type
      properties:
        action_type:
          type: string
          enum:
            - ocr_click
            - ocr_input
            - ocr_wait
            - ocr_assert
            - ocr_get_text
            - image_click
            - image_wait
            - image_assert
            - click
            - swipe
            - input
            - press
            - screenshot
            - wait
            - navigate
            - launch_app
            - close_app
          description: "动作类型"
        value:
          type: string
          description: "文字/URL/按键值"
        image_path:
          type: string
          description: "图像模板路径"
        offset:
          type: object
          properties:
            x:
              type: integer
            y:
              type: integer
          description: "点击偏移"
        threshold:
          type: number
          default: 0.8
          description: "图像匹配阈值"
        timeout:
          type: integer
          default: 30000
          description: "超时时间(ms)"
        match_mode:
          type: string
          enum: [exact, fuzzy, contains, regex]
          default: exact
          description: "OCR 匹配模式"
        screenshot:
          type: boolean
          default: false
          description: "是否截图"
        wait:
          type: integer
          description: "等待时间(ms)"
        x:
          type: integer
          description: "X 坐标"
        y:
          type: integer
          description: "Y 坐标"
        end_x:
          type: integer
          description: "终点 X 坐标（滑动）"
        end_y:
          type: integer
          description: "终点 Y 坐标（滑动）"
        direction:
          type: string
          enum: [up, down, left, right]
          description: "滑动方向"
        app_path:
          type: string
          description: "应用路径"
        bundle_id:
          type: string
          description: "iOS Bundle ID"
        package_name:
          type: string
          description: "Android 包名"

    TaskResult:
      type: object
      properties:
        task_id:
          type: string
          description: "任务 ID"
        status:
          type: string
          enum: [pending, running, success, failed, timeout, cancelled]
          description: "任务状态"
        platform:
          type: string
          description: "平台"
        started_at:
          type: string
          format: date-time
          description: "开始时间"
        finished_at:
          type: string
          format: date-time
          description: "结束时间"
        duration_ms:
          type: integer
          description: "执行耗时(ms)"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/ActionResult"
        screenshots:
          type: array
          items:
            $ref: "#/components/schemas/ScreenshotInfo"
        error:
          type: string
          description: "错误信息"
        metadata:
          type: object
          description: "元数据"

    ActionResult:
      type: object
      properties:
        index:
          type: integer
          description: "动作索引"
        action_type:
          type: string
          description: "动作类型"
        status:
          type: string
          enum: [pending, running, success, failed, skipped]
          description: "动作状态"
        duration_ms:
          type: integer
          description: "执行耗时(ms)"
        output:
          type: string
          description: "输出内容"
        error:
          type: string
          description: "错误信息"
        screenshot:
          type: string
          description: "截图（base64 或路径）"

    ScreenshotInfo:
      type: object
      properties:
        name:
          type: string
          description: "截图名称"
        action_index:
          type: integer
          description: "对应动作索引"
        data:
          type: string
          description: "base64 数据"
        path:
          type: string
          description: "文件路径"

    SessionRequest:
      type: object
      required:
        - platform
      properties:
        platform:
          type: string
          enum: [web, android, ios, windows, mac]
          description: "平台"
        device_id:
          type: string
          description: "设备 ID（移动端必填）"
        options:
          type: object
          description: "会话选项"

    ScreenshotRequest:
      type: object
      required:
        - platform
        - session_id
      properties:
        platform:
          type: string
          description: "平台"
        session_id:
          type: string
          description: "会话 ID"